pyproject.toml
[tool.poetry]
name = "core-project"
version = "0.1.0"
description = ""
authors = ["EvansRobbie <evansitoden94@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "core_project"}]

[tool.poetry.dependencies]
python = "^3.9"
djoser = "^2.2.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-environ = "^0.11.2"
social-auth-app-django = "^5.3.0"
django-cors-headers = "^4.2.0"
docutils = "^0.20.1"
coreapi = "^2.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
